+++ src/expandableCalendar/style.js
@@ -85,6 +85,8 @@ export default function styleConstructor(theme = {}) {
     week: {
       marginTop: 7,
       marginBottom: 7,
+      paddingRight: 15,
+      paddingLeft: 15,
       flexDirection: 'row',
       justifyContent: 'space-around'
     },